// packages/builder/webpack-builder/src/types/config/rem.ts
export type PxToRemOptions = Partial<{
  /** 默认取 RemOptions['rootFontSize'] */
  rootValue: number;
  /** 精确位数 */
  unitPrecision: number;
  /** 支持转换的 css 属性，默认 ['*'] */
  propList: Array<string>;
  selectorBlackList: Array<string>;
  replace: boolean;
  mediaQuery: boolean;
  minPixelValue: number;
  exclude: string | RegExp | ((filePath: string) => boolean);
}>;

export type RemOptions = Partial<{
  /** runtime 选项 */
  /** 是否在 html 模版中注入运行时代码。默认 true */
  enableRuntime: boolean;
  screenWidth: number;
  rootFontSize: number;
  maxRootFontSize: number;
  widthQueryKey: string;
  /** 不进行调整的 entry */
  excludeEntries: Array<string>;
  /** 横屏时使用 height 计算 rem。默认 false */
  supportLandscape: boolean;
  /**
   * 超过 maxRootFontSize 时，是否使用 rootFontSize。
   * 场景：rem 在 pc 上的尺寸计算正常
   */
  useRootFontSizeBeyondMax: boolean;

  /**
   * css 选项
   * 对应 https://github.com/cuth/postcss-pxtorem#options 选项
   */
  pxtorem: PxToRemOptions;
}>;